// src/app/models/contacto.ts
export class Contacto {
  // export interface Articulo {
    $id: number;
    FechaNacimiento: string;
    IdCategoria: number;
    IdContacto: number;
    Nombre: string;
    Telefono : number;
    // CodigoDeBarra: string;
    // //ArticulosFamilia: ArticuloFamilia;
    // Stock: number;
    // Activo: boolean;
  }